+++ app/models/supply_teachers/rm6238/journey/fixed_term_results.rb
@@ -2,6 +2,8 @@ module SupplyTeachers
   module RM6238
     class Journey::FixedTermResults < SupplyTeachers::Journey::FixedTermResults
       include Journey::Results
+
+      POSITION_ID = 'RM6238.1.11'.freeze
     end
   end
 end